# backend/openapi.yaml
openapi: 3.0.3
info:
  title: QM Documents Backend API
  version: 1.0.0
  description: |
    Document template backend for handbook-scoped uploads, placeholder extraction, rendering, and AI rewrite.

    ## Authentication
    **No authentication required.** All endpoints are open.

    ## Scope
    All documents are scoped by `handbook_id`.
paths:
...
  /api/v1/documents:
    get:
      operationId: v1_documents_retrieve
      tags:
      - v1
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          description: No response body
...
  /api/v1/files/tree:
    get:
      operationId: v1_files_tree_retrieve
      tags:
      - v1
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
...